requirements.txt
--find-links https://data.pyg.org/whl/torch-2.2.0+cu118.html
pyg_lib 
torch_scatter 
torch_sparse 
torch_cluster 
torch_spline_conv
datatable==1.1.0